move generator to /ai wip
