+++ packages/nx/src/ai/set-up-ai-agents/get-agent-rules.ts
