+++ packages/nx/src/ai/set-up-ai-agents/schema.d.ts
