+++ packages/nx/src/ai/set-up-ai-agents/set-up-ai-agents.spec.ts
