+++ packages/nx/src/index.ts
@@ -1,2 +1,2 @@
 export { initTasksRunner } from './tasks-runner/init-tasks-runner';
-export { setupAiAgentsGenerator } from './command-line/ai-agent-setup/set-up-ai-agents/set-up-ai-agents';
+export { setupAiAgentsGenerator } from './ai/set-up-ai-agents/set-up-ai-agents';
